IOProxy support for IFF files
